pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hitresearch-sim"
version = "0.1.0"
description = "Headless Isaac Sim + Pegasus data platform for GNSS-denied forest UAV simulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "opencv-python>=4.8",
  "pyyaml>=6.0",
  "pandas>=2.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
